meercat/requirements.txt
arrow==1.2.2
asgiref==3.5.0
astroid==2.11.5
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dill==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
docstring_parser==0.14.1
fonttools==4.33.3
fuzzywuzzy==0.18.0
idna==3.3
isort==5.10.1
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
matplotlib==3.5.2
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.8.0
pylint==2.13.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
textdistance==4.2.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
